[YUNIKORN-108] Display "other" resource types (#127)
Only memory and vcore were displayed on the web UI even if more resource
types were defined. Add dynamic display of different types of resources
based on what is defined on the object.

Set n/a for the value if the type is defined but not a value is given,
i.e. if a resource is empty.

Closes: #127

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>