[YUNIKORN-3076] WebUI fails to load apps that are in 'New' (#233)
When an application is reported to be in a 'New' state by the scheduler, it
does not have a stateLog object. On the web UI side, there are currently no
checks to handle the absence of the stateLog object. As a result, when the
application list contains an application in the 'New' state and the
stateLog is missing, the code attempts to access this.stateLog, leading to
a failure and preventing the page from loading any applications.

Closes: #233

Signed-off-by: Wilfred Spiegelenburg <wilfreds@apache.org>
(cherry picked from commit 541135b7f28c9e63155d76ec730bfd36dcafa27b)